Rename tests to match directories and clean imports
